Fix RacerGetPrefixCol
`racer prefix` returns the byte position, not the column number.

I also defined b:tmpfname in RacerGetExpCompletions. The function relied
on this variable being defined elsewhere, which works, but is dangerous.

I don't know if this was a breaking change in racer. If this was a
breaking change in racer, should we care about backward compatibility?

Fix #68.